refactor(event-stream): extract WriteBacklog to backlog.rs
Pure code motion for #6235. Move the cursor-backed pending socket-write
backlog and its cap out of mod.rs into a dedicated backlog.rs submodule.

Moved verbatim (no logic or signature change): the WRITE_BACKLOG_MAX_BYTES
cap and the WriteBacklog struct with its geometric-compaction impl. All
#[inline] attributes (pending_len, is_empty, pending, compact_if_needed)
are preserved exactly.

The struct, the methods the connected loop / channel drain call
(with_capacity, pending_len, is_empty, pending, extend_from_slice,
advance), and the test-only compacted_bytes field are widened from
private to pub(super) so the still-in-mod.rs callers and the
write_backlog test suite (which reach them via `use super::*`) resolve
unchanged. mod.rs re-imports WriteBacklog and WRITE_BACKLOG_MAX_BYTES.
cargo build green.
